feat(combat): 重构战斗系统并集成配置驱动的怪物生成
- 更新数据库配置支持 allowPublicKeyRetrieval 并修改默认用户
- 将数值类型从 int 升级为 long 以支持更大数值范围
- 重构 CombatAction 为 Spring 组件并注入 CombatService
- 实现基于配置的怪物生成系统，支持地图怪物权重分配
- 添加战斗奖励计算，支持经验、灵石和物品掉落
- 实现挂机战斗系统，包含奖励计算和掉落处理
- 更新 PlayerService 以支持数据库持久化存储
- 集成 RealmService 与配置系统，实现境界属性加成
- 重构 InventoryService 为 Spring 服务并改进线程安全
- 更新 SkillService 为 Spring 服务并添加默认技能初始化

+++ xiaoyao-common/src/main/java/com/xiaoyao/common/proto/BattleResultResp.java
@@ -18,7 +18,7 @@ public class BattleResultResp {
     /** 获得经验 */
     private long expGained;
     /** 获得灵石 */
-    private int stonesGained;
+    private long stonesGained;
     /** 掉落物品 */
     private List<BagItem> drops;
     /** 玩家剩余血量比例 */
