+++ xiaoyao-common/src/main/java/com/xiaoyao/common/proto/BreakthroughResp.java
@@ -11,25 +11,25 @@ import lombok.Data;
 @Data
 @ProtobufClass
 public class BreakthroughResp {
-    
+
     /** 是否成功 */
     private boolean success;
-    
+
     /** 新境界ID */
     private int newRealmId;
-    
+
     /** 新境界名称 */
     private String newRealmName;
-    
+
     /** 攻击力提升 */
-    private int atkBonus;
-    
+    private long atkBonus;
+
     /** 防御力提升 */
-    private int defBonus;
-    
+    private long defBonus;
+
     /** 生命值提升 */
-    private int hpBonus;
-    
+    private long hpBonus;
+
     /** 更新后的战斗力 */
     private long combatPower;
 }
